не подключается скрипт statistic, хотя все пути все прописано правильно

+++ src/index.js
@@ -19,7 +19,7 @@ export let weekMonth = weekDate.getMonth() + 1;
 //     INPUT_FORMA,
 // } from './js/constans/constans.js';
 
-
+console.log(55)
 //переменные API
 import {
     PATH_NEWS,
@@ -38,9 +38,10 @@ import NewsApi from './js/modules/NEWSApi.js';
 export const newsApi = new NewsApi();
 import SearchInput from './js/components/SearchInput.js';
 export const searchInput = new SearchInput();
+
 export const INPUT_FORMA = document.forms.Search; //форма поиска
 
-localStorage.setItem('NUMBER_POSITION', JSON.parse(0));
+//localStorage.setItem('NUMBER_POSITION', 0);
 
 
 
@@ -56,39 +57,46 @@ export const newsCard = new NewsCard();
 
 //document.querySelector('.buttom_place_main').classList.remove('buttom_state.disabled')
 
-INPUT_FORMA.addEventListener('submit', function () {
-    event.preventDefault();
-    if (searchInput.validation(INPUT_FORMA.news)) {
-        document.querySelector('.preloader').classList.add('preloader_state_enabled');
-        //обнуляем страницу
-        localStorage.clear();
-        URL_ARRAY.length = 0;
-        document.querySelector('.body').classList.add('body_preloader');
-        document.querySelector('.circle-preloader').classList.add('circle-preloader_state_disabled');
-        newsCardList.renderCard(document.querySelector('.cards'),
-            document.querySelectorAll('.card'));
+if (typeof INPUT_FORMA !== undefined) {
 
-        newsApi.getNews(PATH_NEWS, INPUT_FORMA.news.value, todayDay,
-                nodayMonth, yearToday, weekDay, weekMonth, weekYear)
-            .then((res) => {
-                document.querySelector('.body').classList.remove('body_preloader');
-                document.querySelector('.body').classList.remove('preloader_error-network')
-                document.querySelector('.circle-preloader').classList.remove('circle-preloader_state_disabled');
-                localStorage.setItem(INPUT_FORMA.news.value, JSON.stringify(res));
-                localStorage.setItem('NewsName', INPUT_FORMA.news.value);
-                newsCardList.pushCard(INPUT_FORMA.news.value, 0);
-                //localStorage.setItem('NUMBER_POSITION', 3)
+    INPUT_FORMA.addEventListener('submit', function () {
+        event.preventDefault();
+        if (searchInput.validation(INPUT_FORMA.news)) {
+            localStorage.setItem('NUMBER_POSITION', 0);
+            document.querySelector('.preloader').classList.add('preloader_state_enabled');
+            //обнуляем страницу
+            localStorage.clear();
+            URL_ARRAY.length = 0;
+            document.querySelector('.body').classList.add('body_preloader');
+            document.querySelector('.circle-preloader').classList.add('circle-preloader_state_disabled');
+            newsCardList.renderCard(document.querySelector('.cards'),
+                document.querySelectorAll('.card'));
 
-            })
-            .catch((res) => {
-                console.log(res)
-                document.querySelector('.preloader').classList.remove('preloader_state_enabled');
-                // document.querySelector('.circle-preloader').classList.remove('circle-preloader_state_disabled');
-                // document.querySelector('.body').classList.remove('body_preloader');
-                // document.querySelector('.preloader__error').classList.remove('preloader_error-network');
-            })
-    }
-})
+            newsApi.getNews(PATH_NEWS, INPUT_FORMA.news.value, todayDay,
+                    nodayMonth, yearToday, weekDay, weekMonth, weekYear)
+                .then((res) => {
+                    document.querySelector('.body').classList.remove('body_preloader');
+                    document.querySelector('.body').classList.remove('preloader_error-network')
+                    document.querySelector('.circle-preloader').classList.remove('circle-preloader_state_disabled');
+                    localStorage.setItem(INPUT_FORMA.news.value, JSON.stringify(res));
+                    localStorage.setItem('NewsName', INPUT_FORMA.news.value);
+                    newsCardList.pushCard(INPUT_FORMA.news.value, 0);
+                    //localStorage.setItem('NUMBER_POSITION', 3)
+
+                })
+                .catch((res) => {
+                    console.log(res)
+                    document.querySelector('.preloader').classList.remove('preloader_state_enabled');
+                    // document.querySelector('.circle-preloader').classList.remove('circle-preloader_state_disabled');
+                    // document.querySelector('.body').classList.remove('body_preloader');
+                    // document.querySelector('.preloader__error').classList.remove('preloader_error-network');
+                })
+        }
+    })
+
+}
+
+
 document.querySelector('.buttom_place_main').addEventListener('click', function (event) {
 
     const newsPosition = JSON.parse(localStorage.getItem('NUMBER_POSITION'));
